+++ docker-compose.yaml
@@ -1,0 +1,6 @@
+version: '3'
+services:
+  autokueng-frontend:
+    image: autokueng-frontend:latest
+    ports:
+      - 3000:3000
