+++ pages/services.js
@@ -1,9 +1,23 @@
+import Services from "../components/services/Services"
 
+// get API Data
+export async function getStaticProps() {
+  const res = await fetch('https://jsonplaceholder.typicode.com/posts')
+  const services = await res.json()
+  
+  return {
+    props: {
+      services,
+    },
 
-export default function Fahrzeugpark() {
+    revalidate: 10,
+  }
+}
+
+export default function Fahrzeugpark({services}) {
     return (
       <>
-        <h1>Hello Services.js</h1>
+        <Services services={services} />
       </>
     )
   }
